Work-around problem where expected files are missing and versionTimestamp is NULL.
The build fails on NULL pointer exceptions and similar errors.

The node iteration "next" is assumed by the code.
When the next does not exist and a NULL is returned then a NULL pointer exception occurs.
Fix this by adding a `hasNext()` check and return NULL if there is no "next".
The caller now handles the NULL check by performing the same task as if there is no versioned resource.

The aclFile and its parent aclDirectory may not exist for unknown reasons.
Add check to see if the either of these exists.

In both cases, provide a warning that can be logged for further investigation.

With this change rather than failing on these errors, the warning log is printed but the upgrade process continues on.